Refactor e2e helpers and tests
