+++ e2e/screens/settings/account.spec.js
@@ -1,11 +1,4 @@
-const helpers = require('./../../helpers');
-const faker = require('faker');
-
-const randomCredentials = {
-  email: faker.internet.exampleEmail(),
-  password: faker.internet.password(),
-  syncServerUrl: 'http://syncing-server.dev:3000'
-};
+const helpers = require('../../helpers');
 
 describe('Account section', () => {
   describe('Form', () => {
@@ -42,11 +35,11 @@ describe('Account section', () => {
     });
 
     it('should work when valid data is provided', async () => {
-      await element(by.id('emailField')).typeText(randomCredentials.email);
-      await element(by.id('passwordField')).typeText(randomCredentials.password);
+      await element(by.id('emailField')).typeText(helpers.randomCredentials.email);
+      await element(by.id('passwordField')).typeText(helpers.randomCredentials.password);
       await element(by.id('otherOptionsButton')).tap();
       await element(by.id('syncServerField')).clearText();
-      await element(by.id('syncServerField')).typeText(randomCredentials.syncServerUrl);
+      await element(by.id('syncServerField')).typeText(helpers.randomCredentials.syncServerUrl);
       await element(by.id('registerButton')).tap();
 
       // A confirmation screen is shown after we click the register button...
@@ -54,7 +47,7 @@ describe('Account section', () => {
       await expect(element(by.id('registerConfirmButton'))).toBeVisible();
 
       // Password confirmation is required...
-      await element(by.id('passwordConfirmationField')).typeText(randomCredentials.password);
+      await element(by.id('passwordConfirmationField')).typeText(helpers.randomCredentials.password);
       await element(by.id('registerConfirmButton')).tap();
     });
 
@@ -77,11 +70,11 @@ describe('Account section', () => {
     });
 
     it('should work when valid data is provided', async () => {
-      await element(by.id('emailField')).typeText(randomCredentials.email);
-      await element(by.id('passwordField')).typeText(randomCredentials.password);
+      await element(by.id('emailField')).typeText(helpers.randomCredentials.email);
+      await element(by.id('passwordField')).typeText(helpers.randomCredentials.password);
       await element(by.id('otherOptionsButton')).tap();
       await element(by.id('syncServerField')).clearText();
-      await element(by.id('syncServerField')).typeText(randomCredentials.syncServerUrl);
+      await element(by.id('syncServerField')).typeText(helpers.randomCredentials.syncServerUrl);
       await element(by.id('signInButton')).tap();
     });
   });
